+++ CHANGES.md
@@ -1,3 +1,7 @@
+# 0.46.0 (2017/12/14)
+  * Remove GUI (now available in a different opam package `grew_gui`)
+  * New command line parsing of arguments, see `grew help`
+  
 ## 0.45.1 (2017/12/07)
 Bugfix: [#3](https://gitlab.inria.fr/grew/grew/issues/3) (Bug in GUI with graphviz 2.40)